move getDataType method to utils
